refactor: moving Word.hpp to Compiler/IR/
